Add optional "space-suffix" parameter (#7)
* Add optional "space-suffix" parameter

In order to allow a switch to Cloud.gov's controlled-egress space types,
adding an optional space-suffix parameter allows us to migrate which
space is targeted by applications one at a time. With this change, all
applications will remain in their current targeted spaces until the
deployment ci/cd for the app switches to providing the "space-suffix"
parameter. Targeting the closed-egress space (named dev-closed) will be
done by adding the "-closed" suffix.

This change preserves all previous outputs and should be non-breaking
for all GitHub actions that leverage this workflow. In order to start
using the suffix versions of the workflow, apps can use the new
"target-space" output of this workflow instead of the
"target-environment" output.